requirements.txt
altair
attrs
blinker
cachetools
certifi
charset-normalizer
click
gitdb0
GitPython7
idna
importlib-metadata
Jinja2
jsonschema
jsonschema-specifications.1
markdown-it-py
MarkupSafe
mdurl
numpy
opencv-python
packaging
pandas
Pillow
protobuf
pyarrow
pydeck
Pygments
pytesseract0
python-dateutil
pytz.post1
referencing
requests
rich
rpds-py
six
smmap
streamlit
tenacity
toml
toolz
tornado
typing_extensions
tzdata
tzlocal
urllib3
validators
Wand3
watchdog
zipp
